Massive keys refactor IPFS generates the node's RSA identity and stores it in the config file. This is problematic for us for two reasons. 1) We don't want keys sitting in plaintext in the config file. Ideally they should be encrypted and separated from the config. 2) We would really like to deterministically generate the identity key from a seed so that we derive both our identity key and bitcoin keychain from a single (bip39 mnemonic) seed.
This commit does some heavy refactoring to change how the repo is initialized
and how the config file is built to move the identity key out of the config file
and derive it from a seed.

At present a basic key expansion algorithm is used turn the bip39 seed into a
pseudorandom stream of bytes to be used by the RSA key generation function (in
place of a purely random entropy source). The exspansion algorithm can probably
be swapped out for a more appropriate standard.