Add a custom JSON field inside metadata
After some conversation with @drwasho (http://imgur.com/a/s6Unw) and @SamPatt (http://imgur.com/a/TT1Pm) a while back, I'd like to submit a pull request to add a text field for JSON which allows nodes to add any amount of structured information to their listings. Extensive third party uses of this field might inform future updates to the protocol that weren't on the radar before or up the priority of those that were.

A good example use case is for a listing posted by a handy man in Cambridge that can only service people in this area. He may want to specify some JSON which allows others to know his location like:

```
{
    "location" : {
        "latitude" : "52.2053° N",
        "longitude" : "0.1218° E"
    }
}
```