Refactor images to use multiple hashes
Change the listing protobuf to accept multiple hashes (tiny, small, medium,
large, original) for each image.

Refactor the SetProductImages function to save images in the new sizes and set
it so that it doesn't crop images, only scales them.

Return each hash in response to the image post json api call.

Save the small, medium, and large hashes in the listing index.