Fix bug setting shipping option type in listing
We have jsonpb set to drop default values when marshalling to json. This results
in enum fields being dropped when set to value zero. So to fix this we have
to modify the protobuf to set value zero to NA or something similar.

This commit applies this change to both the shipping option type and shipping
rule type.

Closes #152