Refactor rating key signature
This commit refactors how the vendor signs the buyer's rating key when sending
the OrderConfirmation and OrderFulfillment messages. Now the buyer's rating key,
the moderator's key (where needed) and the slug added to a protobuf object
which is serialized and signed. This makes it easier to include in and validate
the buyer's rating.