Add listing version check when purchasing
Check the version number of the listing we're trying to purchase does not exceed
the highest version number known to the server. If so the vendor is likely
expecting behavior the buyer cannot produce and hence will cause the vendor
to reject the order. In such cases notify the user he must upgrade to purchase
the listing.