Update dependencies to use custom jsonpb
The way the jsonpb package marshaled some fields to JSON created some weird parts
in the contract proto. For example, enum fields with a value of 0 were dropped
which required us to set all enum fields to NA or something similar. This commit
switches to a fork of jsonpb which removes some of the oddities and also refactors
contracts.proto to reflect the improvement.