Refactor contract signatures
Removes the bitcoin signature from the Signatures list and moves it into the
BuyerID and VendorID sections. The bitcoin signature is only needed once and
covering the GUID. This is to prevent an attack where the vendor tries to give
bad data to the moderator to get him to release the funds.  This refactor reduces
the number of signatures that need to be checked and reduces the size of the
contract.

Closes #199