Add waitgroup to dispute processing
There is a small chance that a DISPUATE_UPDATE message could arrive before we
are done processing the DISPUTE_OPEN message. Normall this would cause the
DISPUTE_UPDATE message to be dropped. This commit adds a waitgroup such that
all DISPUTE_OPEN messages must be finished processing before we proceed with
the DISPUTE_UPDATE message.