Change IPFS gateway PUT handling
Currently put is kind of hard to work with and requires many back and forth
messages to upload a directory. Additionally, we don't really want nodes who
are pushing to the gateway to have to upload their entire directory each time as
this wastes bandwidth.

This commit changes the put handler to accept a directory hash. It will then
traverse the graph and download all nodes it doesn't have.