Display when other user is typing in chat
This commit adds a flag to the chat proto that can be used to send a message
to the recipient that the user is typing. This message gets passed to the UI
as a message with an empty body.

The UI can send these typing messages by just sending a chat message with an
empty message body.

These messages do not get saved into the database and only work when both
parties are online.