Add stealth transaction feature
New feature allows bitcoin payments to be sent to an OpenBazaar peerId or handle.
When this type of payment is made the recipient's master public key is looked
up and a chaincode is generated using ECDH with an ephemeral key pair. The coins
get sent to a seemingly random address, not linked to the master public key.
The ephemeral public key is included in op_return and the recipient can derive
the key needed to spend the coins from the same ECDH operation. The transaction
is relayed by the remote peers by pushing a flag byte to op_return which is also
set in the filter.