Save coupon code to database
The listing schema only saves the hash of a discount code in the contract, but
that means that the client would not see the code that created the hash. So
on listing save (POST or PUT) this saves the code in the database and then
loads it and appends it when a GET listing API call is made.