Listen on websocket transport
The ipfs websocket transport allows the browser to connect directly to ipfs nodes for online communication. The ipfs basichost has this transport enabled by default, but we need to add the relevant address to the swarm. This means that Duo browser clients, or anyone else for that matter, can engage in online comms. I am in the process of testing this and have already had a lot of success sending messages from js to an OB node with this websocket address enabled. This will greatly improve the usability of browser clients.

Fortunately, we can enable this transport in two lines!