Add NewAddress wallet function
Orders need to go to a unique address in order for the listener to determine
what order they are for. This would normally be OK except scenarios where someone
has placed an order but hasn't paid (or two orders come in at the same time).

So we need a wallet function that returns a new address each time. Sadly this
can make it such that malicous buyers can place unfunded orders and run up
the size of the lookahead window in the keychain making it unlikely if not impossible
for all transactions will be detected when restoring from seed.

This can be fixed by either adding the lookahead window size to the config (TODO)
and implementing better DoS protection for unfunded orders.