Enable certain /ob/ endpoints in gateway only mode
Some /ob/ endpoints provide additional functionality over the standard ipfs and ipns
gateway endpoints such as sanitization, caching, and async responses.

The ipfs and ipns endpoints are fine for basic gateway browsing but trying to
build an app on top of the gateway would require access to the endpoints.

This commit changes the behavior of the API enabled config option to keep a
few basic endpoints exposed to the internet. The data here should not be anything
that couldn't be fetched via the ipfs and ipns endpoints.