Save node creation date in db
The spvwallet lib was refactored to add a creation date to the config. The
wallet uses this date to decide what checkpoint to use when initializing the
header db. This commit saves the node creation date to the db and uses it on
startup. It also allows the date to be entered via the command line when running
Init.