Fix bug in IPFS GetHash
GetHash is designed to return the hash of the dag given the bytes. Ideally we would
just calculate it in memory but I can't figure out how to do it now that we
switched to CidV1. For now, we'll just write to a temp file then fetch the hash
like we do for files.

This is something that should be improved in the future.