Update QA tests to check for confirmed payouts
The QA package was mostly only checking that the wallet recorded an unconfirmed
payment and was missing the case where the payment was rejected by the bitcoin
network and failed to confirmed.

This commit updates the package to check the confirmed balance whenever checking
to make sure the funds were received correctly.