Work around windows add file bug
The IPFS add file command randomly fails sometimes on windows. There is an issue
filed in IPFS for it. Seems like it could be related to the indexer.

This commit adds a couple retries to the seed node function before it gives up
and returns an error. Hopefully this will prevent API calls from returning an
error.