Sanitize critical listings fields
Certain fields we do a string comparison on when purchasing the listing and if
the user enters special characters it might cause those fields to get sanitized
when sending the listing to the UI, thus causing the string comparison to fail
when they are returned on purchase. This commit runs the sanitizer on those fields
when saving the listing so that the UI does not get different data.