REFACTOR: Posts
Remove the 'reference' object and replace with 'tags'. Tags is an array of strings that the user can attach to the post.

The idea is that users can fill it with a peerId, slug, location, or a typical social post-style hashtag.

Also remove some superfluous data mistakenly added to the contracts.proto