Add RepublishInterval config option
Adds a new config file option to specify how frequently the user/store data
is republished to IPNS. There is a tradeoff between the desire to have your
data cached on as many nodes as possible and having the latest content (such as
ratings) visible. The config options allows users make this tradeoff according
to their preferences.