Normalize bitcoin cash address when processing dispute
Since the bitcoin cash address format is changing to remove the prefix, it's
possible someone might file a dispute for an order that was created when the
address still had the prefix. This commit adds a check to also check for the
prefix when calculating dispute errors. This prevents it from displaying an
error to the user when there isn't one.

We can remove this check 45 days after release as by then it will not be possible
to file a dispute for an order which used the old address format.