Move zcashd and bitcoind wallets to separate packages
This commit is preparation for the multiwallet integration. Bitcoind and Zcashd
are going to be transitioning to plugins in the future so we don't want to
maintain them in this library.

Also, the ExchangeRate interface has move to the wallet-interface package so this
commit removes it from the bitcoin package here as well.