Add spend all functionality to all wallets
This commit updates the wallet-interface, multiwallet, spvwallet, and
bch spvwallet to add a spendAll bool to the spend interface method that will
allow us to sweep all funds from the wallet. Currently to remove all funds you
need to try to guess what fee the wallet will use and substract that from the
amount you're trying to spend. The spendAll option will subtract the fee from
the amount you're sending rather than adding it. Improving the UX in this regard.

A new bool is added to the spend POST JSON body.

closes #929