Close stream to peer when the peer is banned
The ban manager only checks to see if a peer is banned if that peer tries to
open a new stream. If there is an existing stream to that peer the messages
would otherwise still get through. This commit changes the behavior to close
the streams to any banned peers which should have the effect of preventing
further communication from them.