[#1169] Make transaction and error handling self-contained in repo/db/chat
Chat().Put() has complicated transaction handling which doesn't
attempt to rollback or commit until the very end of large branch in two
paths of logic (once which handles updating a specific messageID and
another more general update). These should probably be extracted, but
this consolidation of logic is a good half-step.