Fix bug when using ExchangeRate().CoinsPerUnit() API
CoinsPerUnit() returns the actual number of coins and wasn't being
calculated properly to return the number of places the unit may be
divided. This corrects the formula to convert the actual number of coins
into the expected divisibility. (Ex: For a unit of currency to be
divisible into 10,000,000 parts, we would expect the divisibility to be
8 as there there would be 8 zeros of precision when represented as the
float of the currency unit.)