Fix double amount in payment record
This code was trying to avoid an edge case where multiple outputs in the
same transaction paid the payment address but this should already have
been accounted for when the record was saved in the db and the values
already aggregated. Thus this was causing incorrect amounts to be saved.
closes #1499