[#1821] Change order of operations in PATCHSettings
New order of operations otimisizes for the least amount of collateral
damage to the network in the event of a failure at any step. Starting
with the most internal checks (validation), then persisting locally, and
then updating the network with side-effects as the last step.

Also ensuring that we properly fail at any given step and returning as
early as possible.