[#1821] SetModeratorsOnListing to operate on protobuf up to Signature
Prefer the protobuf interface for manipulating SignedListing since the
patterns and promises around repo.SignedListing are not so clear when it
comes to signing a listing. Used the existing
NewSignedListingFromProtobuf helper to ensure construction-time promises
are met for the SignedListing object. Expect to reconcile existing
protobuf usage once the repo.SignedListing behavior can be better
protected and refactored into simpler terms.