[#1904] Change repo.Profile truth to use protobuf
repo.Profile had repo structs to mirror each dependent struct in the pb.Profile
which made managing internal state very messy when it came to
serializing to/from protobuf-compatible JSON. This was fine when
repo.Profile was just focused on handling the moderator fee, but I saw
how the scope was expanding and the complexity wasn't worth the benefit
of repo-mirrored-structs which weren't even being used by the consuming
code. Note the public interface of repo.Profile changes significantly
with minimal impact to its usage in core/profile.go.