(*repo.CurrencyValue).ConvertTo accepts a *repo.CurrencyConverter
Isolated the conversion behavior within ConvertTo to live within
repo.CurrencyConverter. It will be provided with a rater interface with
a single method requirement that is satisfied by a subset of today's
wallet-interface.ExchangeRates interface. This change eventually leads
to a wallet-independent exchange rate service which can be separated
later.